# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TES SWAGGER ULBI",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/HammiAhlan",
            "email": "714220062@std.ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "ws-allan2024-0d01e8eb9e77.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/presensi": {
            "get": {
                "description": "Mengambil semua data presensi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Presensi"
                ],
                "summary": "Get All Data Presensi.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Presensi"
                        }
                    }
                }
            }
        },
        "/presensi/{id}": {
            "get": {
                "description": "Ambil per ID data presensi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Presensi"
                ],
                "summary": "Get By ID Data Presensi.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Presensi"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.JamKerja": {
            "type": "object",
            "properties": {
                "durasi": {
                    "type": "integer"
                },
                "gmt": {
                    "type": "integer"
                },
                "hari": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "jam_keluar": {
                    "type": "string"
                },
                "jam_masuk": {
                    "type": "string"
                },
                "piket_tim": {
                    "type": "string"
                },
                "shift": {
                    "type": "integer"
                }
            }
        },
        "controller.Karyawan": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "hari_kerja": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "jabatan": {
                    "type": "string"
                },
                "jam_kerja": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.JamKerja"
                    }
                },
                "nama": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "controller.Presensi": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "biodata": {
                    "$ref": "#/definitions/controller.Karyawan"
                },
                "checkin": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        }
    }
}